q1
Fixed float swap function to have the int temp into float temp

+++ q1.cpp
@@ -37,7 +37,7 @@ void swapTwoValues(int &n1, int &n2) {
 }
 
 void swapTwoValues(float &f1, float &f2) {
-	int temp = f1;
+	float temp = f1;
 	f1 = f2;
 	f2 = temp;
 }
